Organise scripts into packages
